dev: cat-common: Increase size of offset fields

+++ src/dev/cat_common.hh
@@ -13,11 +13,11 @@
  * +=======+====================+
  * | 11000 |      ST_ADDR       |
  * +=======+======+======+======+
- * | 11001 |  ND  | STR2 | OFS2 |
+ * | 11001 | STR2 |    OFS2     |   STR2: 16bit, OFS2: 32bit
  * +=======+======+======+======+
- * | 11010 |  ND  | STR3 | OFS3 |
+ * | 11010 | STR3 |    OFS3     |   STR3: 16bit, OFS3: 32bit
  * +=======+======+======+======+
- * | 11011 |  ND  | OOFS | OLEN |
+ * | 11011 |    OOFS     | OLEN |   OOFS: 32bit, OLEN: 16bit
  * +=======+======+======+======+
  * | 11100 | STR1 | MODE | LEN  |
  * +=======+======+======+======+
@@ -60,10 +60,10 @@ struct params_t
     Addr     start_addr = 0;          // Start address
     int16_t  seq_stride = 0;          // Sequential stride
     int16_t  iv2_stride = 0;          // Interleave 2 stride
-    int16_t  iv2_offset = 0;          // Interleave 2 offset
+    int32_t  iv2_offset = 0;          // Interleave 2 offset
     int16_t  iv3_stride = 0;          // Interleave 3 stride
-    int16_t  iv3_offset = 0;          // Interleave 3 offset
-    int16_t  ol_offset  = 0;          // Outer loop offset
+    int32_t  iv3_offset = 0;          // Interleave 3 offset
+    int32_t  ol_offset  = 0;          // Outer loop offset
     uint16_t ol_length  = 0;          // Outer loop length
     uint16_t length     = 0;          // Interval length
     Addr     tr_b_addr  = 0;          // Translated base address
